requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiokafka==0.7.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorlog==6.7.0
croniter==1.3.7
docopt==0.6.2
faust-streaming==0.9.0
frozenlist==1.3.1
gevent==22.10.1
greenlet==1.1.3.post0
idna==3.4
intervaltree==3.1.0
kafka-python==2.0.2
mode-streaming==0.2.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.23.4
opentracing==2.4.0
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sortedcontainers==2.4.0
terminaltables==3.1.10
urllib3==1.26.12
venusian==3.0.0
websocket==0.2.1
websocket-client==1.4.1
yarg==0.1.9
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.0
